WTA #79: updated dto for ExtractedDataDTO
